Added cast to (long) for conversion from windows time format - to shut up warnings from compiler.

+++ src/gettimeofday.c
@@ -9,7 +9,7 @@ int gettimeofday(struct timeval *tp, void *tz)
 	struct _timeb timebuffer;   
 
 	_ftime( &timebuffer );
-	tp->tv_sec  = timebuffer.time;
+	tp->tv_sec  = (long)timebuffer.time;
 	tp->tv_usec = timebuffer.millitm * 1000;
 	return 0;
 }
